requirements.txt
asgiref==3.8.1
Django==5.2.3
sqlparse==0.5.3
tzdata==2025.2
supabase
python-dotenv
